Move Qwen replay trace helpers out of main
